// interfaces/new.model.interface.ts
export interface NewModelInterface {
  id: number;
  viewsNumber?: number;
  readingTime?: number;
  urlVideo?: null;
  videoDescription?: null;
  appearsHome?: number;
  imageDescription?: null;
  imageAlt?: null;
  urlImage?: string;
  urlImageMobile?: string;
  urlImageThumbnail?: string;
  metaDescription?: string;
  realAuthor?: string;
  disclaimerNotice?: string;
  realDate?: null;
  especialDate?: null;
  authorId?: number;
  statusId?: number;
  categoryId?: number;
  apiCryptoId?: number | null
  languageId?: number
  subcategoryId?: number | null
  createdAt?: any;
  updatedAt?: any;
  author?: any;
  category?: any;
  subcategory?: any;
  newTag?: any;
  newTranslation?: any;
  random?: any;
  slug?: string;
}